+++ CHANGELOG.md
@@ -12,6 +12,24 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.11.1] - 2020-06-10
+### Fixed
+- Client no longer gives an error on using a token instead of an API key.
+
+## [0.11.0] - 2020-06-08
+### Changed
+- Removed FunctionCallResponse class. Functions `FunctionCallsAPI.get_response()` and `FunctionCall.get_response()` now returns the actual function response (without being wrapped in an object with `call_id`, `function_id` and `response`).
+
+## [0.10.0] - 2020-06-04
+### Changed
+- POST endpoints such as search and list will be retried in experimetal endpoints, as they are in v1.
+
+## [0.9.1] - 2020-06-04
+
+### Changed
+- Updated documentation for functions.
+
+
 ## [0.9.0] - 2020-05-29
 ### Changed
 - Function response is no longer a property of the `FunctionCall` class. Instead, the response can be retrieved by the methods `FunctionCallsAPI.get_response()` or `FunctionCall.get_response()`.